% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{apply_list}
\alias{apply_list}
\title{apply list input}
\usage{
apply_list(input, worker)
}
\arguments{
\item{input}{a list}

\item{worker}{a worker}
}
\description{
apply list input
}
\examples{
cutter = worker()
apply_list(list("this is test", "that is not test"), cutter)
}

