% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{DICTPATH}
\alias{DICTPATH}
\alias{HMMPATH}
\alias{IDFPATH}
\alias{STOPPATH}
\alias{USERPATH}
\title{The path of dictionary}
\format{character}
\usage{
DICTPATH

HMMPATH

USERPATH

IDFPATH

STOPPATH
}
\description{
The path of dictionary, and it is used by segmentation and other
function.
}
\keyword{datasets}

