% Generated by roxygen2 (4.0.2): do not edit by hand
\name{distance}
\alias{distance}
\title{Hamming distance of words}
\usage{
distance(codel, coder, jiebar)
}
\arguments{
\item{codel}{A Chinese sentence or the path of a text file.}

\item{coder}{A Chinese sentence or the path of a text file.}

\item{jiebar}{jiebaR Worker.}
}
\description{
The function uses Simhash worker to do keyword extraction worker and find
the keywords, then computes Hamming distance of them.
}
\details{
There is a symbol \code{==} for this function.
}
\examples{
\donttest{
### Simhash
words = "hello world"
simhasher = worker("simhash",topn=1)
simhasher <= words
simhasher == ("hello world" ~ "hello world!")
}
}
\author{
Qin Wenfeng
}
\references{
\url{http://en.wikipedia.org/wiki/Hamming_distance}
}
\seealso{
\code{\link{==.simhash}} \code{\link{worker}}
}

