% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2.R
\name{download_thulac_model}
\alias{download_thulac_model}
\title{Download THULAC Model}
\usage{

  download_thulac_model(model_url = "http://thulac.thunlp.org/source/Models_v1_v2.zip")
}
\description{
Download THULAC Model
}
\examples{

\dontrun{
download_thulac_model()
}
}

