% Generated by roxygen2 (4.0.2): do not edit by hand
\name{==.simhash}
\alias{==.simhash}
\title{Distance symbol}
\usage{
\method{==}{simhash}(jiebar, formula)
}
\arguments{
\item{jiebar}{jiebaR Worker.}

\item{formula}{Two Chinese sentence or the path of text file.}
}
\description{
Distance symbol to compute distance.
}
\examples{
\donttest{
test1 = worker("simhash",topn=1)
test1 == ( "hello world" ~ "hello world")}
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}

