% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dict_tools.R
\name{file_coding}
\alias{file_coding}
\alias{filecoding}
\title{Files encoding detection}
\usage{
file_coding(file)

filecoding(file)
}
\arguments{
\item{file}{A file path.}
}
\value{
The encoding of file
}
\description{
This function detects the encoding of input files. 
You can also check encoding with checkenc package which is on GitHub.
}
\details{
This function will choose the most likely encoding, and it will be 
more stable for a large input text file.
}
\references{
\url{https://github.com/adah1972/tellenc}
}
\seealso{
\url{https://github.com/qinwf/checkenc}
}
\author{
Wu Yongwei, Qin wenfeng
}
