% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/dict_tools.R
\name{filecoding}
\alias{file_coding}
\alias{filecoding}
\title{Files encoding detection}
\usage{
filecoding(file)

file_coding(file)
}
\arguments{
\item{file}{A file path.}
}
\value{
The encoding of file
}
\description{
The function detect the encoding of input files.
}
\author{
Wu Yongwei, Qin wenfeng
}
\references{
\url{https://github.com/adah1972/tellenc}
}

