% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_hardware_concurrency}
\alias{get_hardware_concurrency}
\title{Get number of concurrent threads}
\usage{
get_hardware_concurrency()
}
\value{
number of concurrent threads
}
\description{
Get number of concurrent threads supported by the implementation in C++11.
}
\references{
\url{http://en.cppreference.com/w/cpp/thread/thread/hardware_concurrency}
}

