% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_idf.R
\name{get_idf}
\alias{get_idf}
\title{generate IDF dict}
\usage{
get_idf(x, stop_word = STOPPATH, path = NULL)
}
\arguments{
\item{x}{a list of character}

\item{stop_word}{stopword path}

\item{path}{output path}
}
\description{
If path is not NULL, it will write the result to the path
}

