% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/v2.R
\name{get_thulacr_path}
\alias{get_thulacr_path}
\title{Get jiebaR Package Path}
\usage{
get_thulacr_path(index = 1)
}
\arguments{
\item{index}{If there are multiple jiebaR installed in the system, select one of them to use.}
}
\description{
Get jiebaR Package Path
}
\examples{

\dontrun{
get_thulac_path()
}
}

