% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tuple.R
\name{get_tuple}
\alias{get_tuple}
\title{get tuple from the segmentation result}
\usage{
get_tuple(x, size = 2, dataframe = T)
}
\arguments{
\item{x}{a character vector or list}

\item{size}{a integer >= 2}

\item{dataframe}{return data.frame}
}
\description{
get tuple from the segmentation result
}
\examples{
get_tuple(c("sd","sd","sd","rd"),2)
}
