% Generated by roxygen2 (4.0.2): do not edit by hand
\name{keywords}
\alias{keywords}
\title{Keyword extraction}
\usage{
keywords(code, jiebar)
}
\arguments{
\item{code}{A Chinese sentence or the path of a text file.}

\item{jiebar}{jiebaR Worker.}
}
\value{
a vector of keywords with weight.
}
\description{
Keyword Extraction worker use MixSegment model to cut word and use
TF-IDF algorithm to find the keywords.  \code{dict} , \code{hmm},
\code{idf}, \code{stop_word} and \code{topn} should be provided when initializing
jiebaR worker.
}
\details{
There is a symbol \code{<=} for this function.
}
\examples{
\donttest{
### Keyword Extraction
keys = worker("keywords", topn = 1)
keys <= "words of fun"}
}
\author{
Qin Wenfeng
}
\references{
\url{http://en.wikipedia.org/wiki/Tf-idf}
}
\seealso{
\code{\link{<=.keywords}} \code{\link{worker}}
}

