% Generated by roxygen2 (4.0.2): do not edit by hand
\name{<=.keywords}
\alias{<=.keywords}
\alias{[.keywords}
\title{Keywords symbol}
\usage{
\method{<=}{keywords}(jiebar, code)

\method{[}{keywords}(jiebar, code)
}
\arguments{
\item{jiebar}{jiebaR Worker.}

\item{code}{A Chinese sentence or the path of a text file.}
}
\description{
Keywords symbol to find keywords.
}
\examples{
\donttest{
words = "hello world"
test1 = worker("keywords",topn=1)
test1 <= words}
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}

