% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{<=.qseg}
\alias{<=.qseg}
\alias{qseg}
\title{Quick mode symbol}
\format{qseg an environment}
\usage{
\method{<=}{qseg}(qseg, code)

qseg
}
\arguments{
\item{qseg}{a qseg object.}

\item{code}{a string}
}
\description{
Quick mode symbol to do segmentation, keyword extraction
and speech tagging. This symbol will initialize a \code{quick_worker} when it is
first called, and will do segmentation or other work immediately.
}
\details{
You can reset the default model setting by \code{$}, and it will change the
default setting when the next time you use quick mode. If you only want to
change the parameter temporary, you can reset the settings of \code{quick_worker}.

\code{get_qsegmodel}, \code{set_qsegmodel},and \code{reset_qsegmodel} are
also available for setting quick mode settings.
}
\examples{
\donttest{
qseg <= "This is test"
qseg <= "This is the second test"
}

\dontrun{
qseg <= "This is test"
qseg$detect = T
qseg
get_qsegmodel()
}
}
\author{
Qin Wenfeng
}
\seealso{
\code{\link{set_qsegmodel}} \code{\link{worker}}
}
\keyword{datasets}

