% Generated by roxygen2 (4.0.2): do not edit by hand
\name{<=.simhash}
\alias{<=.simhash}
\alias{[.simhash}
\title{Simhash symbol}
\usage{
\method{<=}{simhash}(jiebar, code)

\method{[}{simhash}(jiebar, code)
}
\arguments{
\item{jiebar}{jiebaR Worker.}

\item{code}{A Chinese sentence or the path of a text file.}
}
\description{
Simhash symbol to compute simhash.
}
\examples{
\donttest{
words = "hello world"
test1 = worker("simhash",topn=1)
test1 <= words}
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}

