% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overload.R, R/print.R
\name{print.inv}
\alias{print.inv}
\alias{print.jieba}
\alias{print.simhash}
\alias{print.keywords}
\alias{print.qseg}
\title{Print worker settings}
\usage{
\method{print}{inv}(x, ...)

\method{print}{jieba}(x, ...)

\method{print}{simhash}(x, ...)

\method{print}{keywords}(x, ...)

\method{print}{qseg}(x, ...)
}
\arguments{
\item{x}{The jiebaR Worker.}

\item{...}{Other arguments.}
}
\description{
These functoins print the worker settings.
}
\author{
Qin Wenfeng
}
