% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_func.R
\name{query_threshold}
\alias{query_threshold}
\title{Set query threshold}
\usage{
query_threshold(worker, number)
}
\arguments{
\item{worker}{a jieba worker}

\item{number}{the query threshold}
}
\description{
Depreciated. This function will be remove in the next version.
}

