% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/quick.R
\name{get_qsegmodel}
\alias{get_qsegmodel}
\alias{reset_qsegmodel}
\alias{set_qsegmodel}
\title{Set quick mode model}
\usage{
get_qsegmodel()

set_qsegmodel(qsegmodel)

reset_qsegmodel()
}
\arguments{
\item{qsegmodel}{a list which has the same structure as the return value of get_qsegmodel}
}
\description{
These function can get and modify quick mode model. \code{get_qsegmodel} returns
the default model parameters. \code{set_qsegmodel} can modify quick mode model
using a list, which has the same structure as the return value of get_qsegmodel.
\code{reset_qsegmodel} can reset the default model to origin \code{jiebaR} default
model.
}
\examples{
\dontrun{
qseg <= "This is test"
qseg <= "This is the second test"
}

\dontrun{
qseg <= "This is test"
qseg$detect = T
qseg
get_qsegmodel()
model = get_qsegmodel()
model$detect = F
set_qsegmodel(model)
reset_qsegmodel()
}
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}
\seealso{
\code{\link{qseg}} \code{\link{worker}}
}

