% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagger.R
\name{tagging}
\alias{tagging}
\title{Speech Tagging}
\usage{
tagging(code, jiebar)
}
\arguments{
\item{code}{a Chinese sentence or the path of a text file}

\item{jiebar}{jiebaR Worker}
}
\description{
The function uses Speech Tagging worker to cut word and 
tags each word after segmentation using labels compatible with 
ictclas.  \code{dict} 
\code{hmm} and \code{user} should be provided when initializing 
jiebaR worker.
}
\details{
There is a symbol \code{<=} for this function.
}
\examples{
\dontrun{
words = "hello world"

### Speech Tagging 
tagger = worker("tag")
tagger <= words
}
}
\references{
The ictclas speech tag : \url{http://t.cn/RAEj7e1}
}
\seealso{
\code{\link{<=.tagger}} \code{\link{worker}}
}
\author{
Qin Wenfeng
}
