% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagger.R
\name{vector_tag}
\alias{vector_tag}
\title{Tag the a character vector}
\usage{
vector_tag(string, jiebar)
}
\arguments{
\item{string}{a character vector of segmented words.}

\item{jiebar}{jiebaR Worker.}
}
\description{
Tag the a character vector
}
\examples{
\dontrun{
cc = worker()
(res = cc["this is test"])
vector_tag(res, cc)
}

}
