% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tools.R
\name{edit_dict}
\alias{EditDict}
\alias{edit_dict}
\title{Edit default user dictionary}
\usage{
edit_dict(name = "user")

EditDict(name = "user")
}
\arguments{
\item{name}{the name of dictionary including \code{user}, \code{system},
\code{stop_word}.}
}
\description{
Edit the default user dictionary.
}
\details{
There are three column in the system dictionary. The first column is the word,
and the second column is the frequency of word. The third column is
speech tag using labels compatible with ictclas.

There are two column in the user dictionary. The first column is the word,
and the second column is speech tag using labels compatible with ictclas.
Frequency of every word in the user dictionary will be the maximum number of
the system dictionary. If you want to provide the frequency of a new word,
you can put it in the system dictionary.
}
\author{
Qin Wenfeng
}
\references{
The ictclas speech tag : \url{http://t.cn/RAEj7e1}
}

