% Generated by roxygen2 (4.0.2): do not edit by hand
\name{edit_dict}
\alias{EditDict}
\alias{edit_dict}
\title{Edit default user dictionary}
\usage{
edit_dict()

EditDict()
}
\description{
Edit the default user dictionary.
}
\details{
There are three column in the dictionary. The first column is the word,
and the second column is the frequency of words. The third column is
speech tag using labels compatible with ictclas.
}
\author{
Qin Wenfeng
}
\references{
The ictclas speech tag : \url{http://t.cn/8FdDD3I}
}

