% Generated by roxygen2 (4.0.2): do not edit by hand
\name{filecoding}
\alias{filecoding}
\title{Files encoding detection}
\usage{
filecoding(file)
}
\arguments{
\item{file}{A file path.}
}
\value{
The encoding of file
}
\description{
The function detect the encoding of input files.
}
\author{
Wu Yongwei, Qin wenfeng
}
\references{
\url{https://github.com/adah1972/tellenc}
}

