% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_segment}
\alias{filter_segment}
\title{Filter segmentation result
This function helps remove some words in the segmentation result.}
\usage{
filter_segment(input, filter_words)
}
\arguments{
\item{input}{a string vector}

\item{filter_words}{a string vector of words to be removed.}
}
\description{
Filter segmentation result
This function helps remove some words in the segmentation result.
}

