% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/words_freq.R
\name{freq}
\alias{freq}
\title{The frequency of words}
\usage{
freq(x)
}
\arguments{
\item{x}{a vector of words}
}
\value{
The frequency of words
}
\description{
This function returns the frequency of words
}
\examples{
freq(c("a","a","c"))
}
\author{
Qin wenfeng
}

