% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_idf.R
\name{get_idf}
\alias{get_idf}
\title{generate IDF dict}
\usage{
get_idf(x, stop_word = STOPPATH, path = NULL)
}
\arguments{
\item{x}{a list of character}

\item{stop_word}{stopword path}

\item{path}{output path}
}
\value{
a data.frame or a file
}
\description{
Generate IDF dict from a list of documents.
}
\details{
Input list contains multiple character vectors with words, 
and each vector represents a document.

Stop words will be removed from the result.

If path is not NULL, it will write the result to the path.
}
\examples{
get_idf(list(c("abc","def"),c("abc"," ")))
}
\seealso{
https://en.wikipedia.org/wiki/Tf-idf#Inverse_document_frequency_2
}
