% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overload.R
\name{<=.keywords}
\alias{<=.keywords}
\alias{[.keywords}
\title{Keywords symbol}
\usage{
\method{<=}{keywords}(jiebar, code)

\method{[}{keywords}(jiebar, code)
}
\arguments{
\item{jiebar}{jiebaR Worker.}

\item{code}{A Chinese sentence or the path of a text file.}
}
\description{
Keywords symbol to find keywords.
}
\examples{
\dontrun{
words = "hello world"
test1 = worker("keywords",topn=1)
test1 <= words}
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}

