% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/overload.R
\name{<=.segment}
\alias{<=.segment}
\alias{[.segment}
\title{Text segmentation symbol}
\usage{
\method{<=}{segment}(jiebar, code)

\method{[}{segment}(jiebar, code)
}
\arguments{
\item{jiebar}{jiebaR Worker.}

\item{code}{A Chinese sentence or the path of a text file.}
}
\description{
Text segmentation symbol to cut words.
}
\examples{
\dontrun{
words = "hello world"
test1 = worker()
test1 <= words}
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}

