% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.hmmseg}
\alias{print.hmmseg}
\alias{print.keywords}
\alias{print.mixseg}
\alias{print.mpseg}
\alias{print.queryseg}
\alias{print.simhash}
\alias{print.tagger}
\title{Print worker settings}
\usage{
\method{print}{hmmseg}(jiebar, ...)

\method{print}{queryseg}(jiebar, ...)

\method{print}{simhash}(jiebar, ...)

\method{print}{keywords}(jiebar, ...)

\method{print}{tagger}(jiebar, ...)

\method{print}{mixseg}(jiebar, ...)

\method{print}{mpseg}(jiebar, ...)
}
\arguments{
\item{jiebar}{The jiebaR Worker.}
}
\description{
The functoin print the worker settings.
}
\author{
Qin Wenfeng <\url{http://qinwenfeng.com}>
}

