% Generated by roxygen2 (4.0.2): do not edit by hand
\name{print.hmmseg}
\alias{print.hmmseg}
\alias{print.keywords}
\alias{print.mixseg}
\alias{print.mpseg}
\alias{print.queryseg}
\alias{print.simhash}
\alias{print.tagger}
\title{Print worker settings}
\usage{
\method{print}{hmmseg}(x, ...)

\method{print}{queryseg}(x, ...)

\method{print}{simhash}(x, ...)

\method{print}{keywords}(x, ...)

\method{print}{tagger}(x, ...)

\method{print}{mixseg}(x, ...)

\method{print}{mpseg}(x, ...)
}
\arguments{
\item{x}{The jiebaR Worker.}

\item{...}{Other arguments.}
}
\description{
The functoin print the worker settings.
}
\author{
Qin Wenfeng
}

