% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/overload.R, R/print.R
\name{print.inv}
\alias{print.hmmseg}
\alias{print.inv}
\alias{print.keywords}
\alias{print.mixseg}
\alias{print.mpseg}
\alias{print.queryseg}
\alias{print.simhash}
\alias{print.tagger}
\title{Print worker settings}
\usage{
\method{print}{inv}(x, ...)

\method{print}{hmmseg}(x, ...)

\method{print}{queryseg}(x, ...)

\method{print}{simhash}(x, ...)

\method{print}{keywords}(x, ...)

\method{print}{tagger}(x, ...)

\method{print}{mixseg}(x, ...)

\method{print}{mpseg}(x, ...)
}
\arguments{
\item{x}{The jiebaR Worker.}

\item{...}{Other arguments.}
}
\description{
These functoins print the worker settings.
}
\author{
Qin Wenfeng
}

