% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dict_tools.R
\name{show_dictpath}
\alias{show_dictpath}
\title{Show default path of dictionaries}
\usage{
show_dictpath()
}
\description{
Show the default dictionaries' path. \code{HMMPATH}, \code{DICTPATH}
, \code{IDFPATH}, \code{STOPPATH} and \code{USERPATH} can be changed 
in default environment.
}
\author{
Qin Wenfeng
}
