% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker_func.R
\name{words_locate}
\alias{words_locate}
\title{Get text location}
\usage{
words_locate(words)
}
\arguments{
\item{words}{a string}
}
\value{
a list with words, start position, and end position
}
\description{
Get text location
}
\examples{
words_locate(c("this","is","a", "test"))

}

